requirements.txt
Pillow
paddlepaddle
paddleocr
requests
shapely
pyclipper
lmdb
